+++ crates/hlight/src/lib.rs
@@ -5,9 +5,11 @@
 
 A library for output syntax highlighting.
 */
+mod color_escape;
 pub mod output;
 mod resource;
 pub mod syntax;
 pub mod theme;
+pub use color_escape::to_ansi_256color;
 
 pub use crate::{output::Highlighter, resource::HighlightResource};